--- backend/app/core/deps/__init__.py
@@ -1,1 +1,0 @@
-from .index import *
