--- backend/app/core/notifications/__init__.py
@@ -1,15 +1,0 @@
-from app.core.notifications.service import NotificationService
-from app.core.notifications.channels import EmailChannel, SlackChannel, WhatsAppChannel
-from app.core.notifications.events import (
-    SendPushNotificationEvent
-)
-from app.core.notifications.templates import TemplateEngine
-
-__all__ = [
-    "NotificationService",
-    "EmailChannel",
-    "SlackChannel",
-    "WhatsAppChannel",
-    "TemplateEngine",
-    "SendPushNotificationEvent"
-]
